+++ src/components/table/table_header_cell.js
@@ -30,13 +30,15 @@ export const EuiTableHeaderCell = ({
   isSortAscending,
   className,
   scope,
+  mobileOptions,
+  // Soon to be deprecated for {...mobileOptions}
   isMobileHeader,
   hideForMobile,
   ...rest
 }) => {
   const classes = classNames('euiTableHeaderCell', className, {
-    'euiTableHeaderCell--isMobileHeader': isMobileHeader,
-    'euiTableHeaderCell--hideForMobile': hideForMobile,
+    'euiTableHeaderCell--hideForDesktop': mobileOptions.only || isMobileHeader,
+    'euiTableHeaderCell--hideForMobile': !mobileOptions.show || hideForMobile,
   });
 
   const contentClasses = classNames('euiTableCellContent', className, {
@@ -111,18 +113,36 @@ EuiTableHeaderCell.propTypes = {
   isSortAscending: PropTypes.bool,
   scope: PropTypes.oneOf(['col', 'row', 'colgroup', 'rowgroup']),
   /**
+   * _DEPRECATED: use `mobileOptions.only = true`_
    * Indicates if the column was created to be the row's heading in mobile view
    * (this column will be hidden at larger screens)
    */
   isMobileHeader: PropTypes.bool,
   /**
+   * _DEPRECATED: use `mobileOptions.show = false`_
    * Indicates if the column should not show for mobile users
    * (typically hidden because a custom mobile header utilizes the column's contents)
    */
   hideForMobile: PropTypes.bool,
+  /**
+   * Mobile options for displaying differently at small screens
+   */
+  mobileOptions: PropTypes.shape({
+    /**
+     * If false, will not render the column at all for mobile
+     */
+    show: PropTypes.bool,
+    /**
+     * Only show for mobile? If true, will not render the column at all for desktop
+     */
+    only: PropTypes.bool,
+  }),
 };
 
 EuiTableHeaderCell.defaultProps = {
   align: LEFT_ALIGNMENT,
   scope: 'col',
+  mobileOptions: {
+    show: true,
+  }
 };